requirements.txt
sentence-transformers==5.0.0
faiss-cpu==1.11.0
requests==2.32.4
numpy==1.26.4